Refactor IMS clients
